Create basic Home page, Feed, and PostCard.
This post creates a basic boilerplate for the home page (rendered by
the `Home`component).

The Home page will be in charge of rendering a list of posts from the
store. This list of posts is called "Feed" and so a `Feed` component is
also created. The Feed component iterates over an array of posts and
renders a PostCard for each one of them.

The PostCard component will be displayed as a compact version of the
posts including only the content of each post (title and body) and will
be used as links to open the detailed view of each post (this view will
include the comments).