Create basic initial redux store
This commit adds the store/ folder and creates a slice for the feed of
posts. Also the Home page is connected to the slice so it can dispatch
an action to fetch posts and populate the feed.

This dispatched action is defined in the slice as a thunk function that
just returns a dummy set of posts after two seconds (to simulate the
response time of the service), but it will be changed by an actual fetch
call in future commits.